+++ src/main/kotlin/com/iplease/server/ip/manage/assign/repository/AssignedIpRepository.kt
@@ -1,6 +1,6 @@
 package com.iplease.server.ip.manage.assign.repository
 
-import com.iplease.server.ip.manage.assign.table.AssignedIpTable
+import com.iplease.server.ip.manage.assign.data.table.AssignedIpTable
 import org.springframework.data.r2dbc.repository.R2dbcRepository
 
 interface AssignedIpRepository: R2dbcRepository<AssignedIpTable, Long>
