+++ src/main/kotlin/com/iplease/server/ip/manage/assign/service/IpAssignService.kt
@@ -1,6 +1,6 @@
 package com.iplease.server.ip.manage.assign.service
 
-import com.iplease.server.ip.manage.assign.dto.AssignedIpDto
+import com.iplease.server.ip.manage.assign.data.dto.AssignedIpDto
 import reactor.core.publisher.Mono
 
 interface IpAssignService {
