+++ src/main/kotlin/com/iplease/server/ip/manage/event/data/type/Event.kt
@@ -1,4 +1,4 @@
-package com.iplease.server.ip.manage.event.type
+package com.iplease.server.ip.manage.event.data.type
 
 enum class Event(
     val routingKey: String
